+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## devel
 
+* add cluster memory and cpu capacity to summary
 * add cluster version to summary
 
 ## 0.2.1